requirements.txt
-f file:///home/tin/.pip_packages
Django==1.6
South==0.8.2
argparse==1.2.1
distribute==0.6.34
django-allauth==0.14.1
django-bootstrap-toolkit==2.15.0
django-tastypie==0.10.0
django-extensions==1.2.5
factory-boy==2.2.1
ipython==1.1.0
oauthlib==0.6.0
python-dateutil==2.2
python-openid==2.2.5
requests==2.0.1
requests-oauthlib==0.4.0
six==1.4.1
wsgiref==0.1.2
